update documentation for getValue/select
`select` elements work with `.getValue()`. This documentation is slightly misleading, as it shows that it would only work for `textarea` or `input`.

As a reference, [this page](http://webdriver.io/api/action/selectByValue.html) in the `selectByValue.js` example on line 6, it shows `.getValue()` being used with a `select` element.

Former-commit-id: 7d3df351fd92be5fb249e5c46b760ef4fd5c9f22