Create selenium-standalone.md
I find it helpful when the documentation does not make assumptions about what the reader knows. In this case, the docs are not clear where the options should be placed. Seeing an example of one of the options in use makes it clear as to where they should go.

Don't make me experiment, just show me how they work.