Support '.es6' files
In our project, we are suffixing all module files targeting ecmascript 6 with '.es6', because of several reasons, mainly because our IDE will interpret '.js' as ecmascript 5 and give many error messages.

I tried to set the pattern to '*.spec.es6' but it would not work. I think that, if I have specified an extension in the pattern, it should be allowed through. But if the extension is a star, or contains a star, you should only pick JS-files, as before.