Webdriver and WebdriverIO Typing (#3218)
## Proposed changes

Adding Typescript type definitions for v5, as well as an automatic generation of WebdriverIO definitions. Also, updating the JS docs to match the current parameters.

## Types of changes

[//]: # (What types of changes does your code introduce to WebdriverIO?)
[//]: # (_Put an `x` in the boxes that apply_)

- [ ] Bugfix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist

[//]: # (_Put an `x` in the boxes that apply. You can also fill these out after creating the PR. If you're unsure about any of them, don't hesitate to ask. We're here to help! This is simply a reminder of what we are going to look for before merging your code._)

- [x] I have read the [CONTRIBUTING](https://github.com/webdriverio/webdriverio/blob/master/CONTRIBUTING.md) doc
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added necessary documentation (if appropriate)

## Further comments

So this is my first time creating type definitions, and I'm sure there are some commands that need to be added. This is building off of @christian-bromann PR [2862](https://github.com/webdriverio/webdriverio/pull/2862)

I have placed definition files in the ```webdriver``` and ```webdriverio``` packages, and set the ```types``` to point to them, so they will be automatically imported with the repo.

I have placed the definition files in my own project to test, so I know that what is there currently works.

### Reviewers: @webdriverio/technical-committee