webdriverio: Allow selectByVisibleText to work with newlines (#3716)
With the addition of #3540 selectByVisibleText will not match options like the following example. This adds back the ```normalize-space``` selection while keeping the current selections.

Although I think this should probably be handled/fixed by the users app, this returns the functionality of how it previously acted. We have quite a few places in our app where we have options formatted like this.

```
$(`.test`).selectByVisibleText(`foo`)

<option>Foo
</option>
```

## Types of changes

- [x] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist

- [x] I have read the [CONTRIBUTING](https://github.com/webdriverio/webdriverio/blob/master/CONTRIBUTING.md) doc
- [x] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added necessary documentation (if appropriate)

### Reviewers: @webdriverio/technical-committee