Print reporters after stdout
Many times I want to see the reporter information (especially from 'spec' reporter) more than I want to see Stdout info. This change will show the reporter details below the stdout in the final output, making it easier to get to. This really helps when there's a lot of stdout in the way of getting to the reporter information.