Appium service (#3717)
## Proposed changes

The [wdio-appium-service](http://github.com/rhysd/wdio-appium-service) doesn't work with wdio v5. It was decided that instead of fixing the community package, it should be absorbed into the wdio monorepo (#3693).

## Types of changes

- [ ] Bugfix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist

- [x] I have read the [CONTRIBUTING](https://github.com/webdriverio/webdriverio/blob/master/CONTRIBUTING.md) doc
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I have added necessary documentation (if appropriate)

## Further comments

When creating this service I took a lot of inspiration from @wdio/selenium-standalone-service and [wdio-appium-service](http://github.com/rhysd/wdio-appium-service).

Most of the functionality is like [wdio-appium-service](http://github.com/rhysd/wdio-appium-service) but adapted for @wdio fx using the @wdio/logger. I also changed some naming and added some comments.

The `getFilePath.js` is a copy from @wdio/selenium-standalone-service. This should maybe be moved to @wdio/utils or something. I've copied the tests for it.

I'm not sure how much we should test in this service. I've created tests for all helper methods and also a test for checking that the Appium service was started. I couldn't get it to succeed all the time. I hope someone can help with this. I can see that the service is started by making a HTTP GET request to `http://127.0.0.1:4723/wd/hub/status`, but in the test I always get a connection refused.
I haven't created a test for the `exitCallback` as I couldn't find a way to trigger it. I think the service needs to be killed before the timeout.

### Reviewers: @webdriverio/technical-committee