checkUnicode should always return an array regardless of isDevTools
There's code relying on the fact that `checkUnicode` returns an array.
For example, `commands/browser/keys.js` assumes this.

The addition of the `isDevTools` flag, introduced in 49eda5119d6b7504d25de1ac17e9b60e6f24f419
messes this up and results in an inconsistent return value.

This was caught because examples for multiple remote browsers failed.

Fixes #5087