devtools: Fix `ignoreDefaultArgs` capability option
The value of `ignoreDefaultArgs` was previously passed directly to `launch`
from the package `chrome-launcher`. However, that function doesn't take the
same parameter, it takes an `ignoreDefaultFlags` boolean parameter. It also
doesn't do any filtering of arguments, it's either just ignoring all default
flags or using all of them.

This changes code in `devtools` to always send `ignoreDefaultFlags: true`. This
should be fine, since we have the default arguments in a list in this project,
and passing all default arguments, together with custom arguments, onwards to
`chrome-launcher`.

To not break previous behavior, I'm also adding these arguments as default
arguments, as they were used from within the default arguments in
`chrome-launcher`, but not in `devtools`:
* `--no-default-browser-check`
* `--force-fieldtrials=*BackgroundTracing/default/`