webdriver: Fix `outputDir` option
There is one log statement that happens above the place where `WDIO_LOG_PATH`
was set, which meant that that log would be logged to stdout, and then
everything else would be sent to file. By setting `WDIO_LOG_PATH` earlier,
everything is sent to file as expected.

There was also a problem with `validateConfig` where the tests were using the
mock file defined inside `packages/wdio-config/tests/__mocks__/@wdio/config.ts`
causing it to get a default value for `outputDir`. By not mocking that module
anymore, we get the real implementation, which will use the values we provide,
and that uncovered a bug in the constants for `webdriver`.

The constant had to be moved from `webdriverio` to `webdriver`, as that seems
to have been missed before. The default value is documented as `null` on the
website, and that has to remain to keep the functionality described (if not set,
the logs are streamed to stdout). However, it seems due to TypeScript types
we're actually using `undefined` in code rather than `null`.