devtools: Add `outputDir` support
This aligns it with the `webdriver` protocol, so that a change of protocol
alone doesn't change where logs go. If `outputDir` is set from a higher level
and passed down to the protocol (`webdriver` or `devtools`), it should be
respected regardless of the protocol, and logs should go to the right file.