Move initial logs from `webdriverio` into driver modules
Since each driver is now responsible for setting up the log path needed for
writing to file, we can't do any logging in `webdriverio` before
`await ProtocolDriver.newSession()` is done, since those logs would then go to
stdout instead of the log file.

By removing this specific log from `webdriverio` and adding it to `webdriver`
and `devtools` we get the same log output, but done to the right place (stdout
or file, depending on `outputDir`).