fix: Remove scripts from log payload  (#6965)
* fix: Remove scripts from log payload 

Added check for `script` property in payload and returned placeholder if it existed. Checks were the same as those for the file property on the `results` object

Resolves issue #6714

* feat: Replaced previous placeholder text

Changed the default placeholder for the script from 'Payload'
to '"Script<base64>"'. I also made the new default value a
constant in case it is needed elsewhere in the file

Resolves Issue: #6921

* test: extended 'transformCommandLogResult'

Extended test to cater for default script placeholder
functionality

Resolves Issue: #6921