(docs): saveScreenshot requires afterTest hook to be async
When requesting a screenshot in the `afterTest` hook, I would sometime
not have any screenshot generated. I eventually found
`browser.saveScreenshot` is an async function and the hook has to be
explicitly defined `async`:

https://github.com/webdriverio/webdriverio/issues/5545#issuecomment-1215272473

Add an example to the `browser.saveScreenshot` documentation showing how
to use it in the `afterTest` hook.

Signed-off-by: Antoine Musso <amusso@wikimedia.org>