Improve automated webdriver setup documentation (#11128)
WHAT?

Improve the automated webdriver setup documentation introduced in `8.14`.

WHY?

While this feature is awesome, I have a test setup environment with no
Internet access. I was confused on how to disable this feature since the
documentation did not cover this topic in depth. After looking into the
source code to figure it out, I decided to pay-it-forward for other
users by updating the documentation accordingly.

Co-authored-by: Christian Bromann <git@bromann.dev>