Fix incorrect browser auto-setup doc
WHAT?

Fix an incorrect statement that says `browserVersion` is used to locate
an existing installation of the browser, when in reality instead it
always trigger a download.

WHY?

When I contributed #11128, I misunderstood the `browserVersion` logic.
This update corrects this misunderstanding.

HOW?

Manually reviewed [packages/wdio-utils/src/driver/utils.ts][1]

[1]: https://github.com/webdriverio/webdriverio/blob/v8.16.7/packages/wdio-utils/src/driver/utils.ts