docs: update autowait limitations example (#11828)
update the numeric index used in the last javascript example on the page to access the array with index 1. the conditional above it returns `false` if the length of the array is not 2. accessing an array at `elms[2]` on a non-sparse array may not be possible in this situation.

I then worked backwards to the previous example, updated the array access there too for continuity.