Update error handling and test logging in allure-reporter
This update modifies the wdio-allure-reporter package to better handle errors during test execution. If a step or substep fails, the error message and stack trace are now recorded for the current test. In addition, conditional logging for failed hooks has been implemented, and changes have been made to ensure the correct number of test results are recorded in suite tests.

Signed-off-by: Lara, Manuel <lararojas.mr@gmail.com>