Fix typo in `addInitScript` example in v9 blog post
In the blogpost announcing Webdriverio v9, the `addInitScript()` example shows a function that takes `callback` as a parameter, but then calls `emit` instead (after looking at the docs, this is clearly supposed to be the parameter named `callback`). I changed the name of the parameter to `emit` to match the examples in the actual docs, although which name to use is pretty arbitrary.