fix(wdio-cucumber-framework): filter cucumberFeaturesWithLineNumbers (webdriverio#13560)
While `_specs` has already been filtered here, `cucumberFeaturesWithLineNumbers`,
which just exists to retain the line numbers, still contains all specs as passed
on the cmdline or in the config. Therefore, just passing these to cucumber will
inevitably execute all feature lines in every worker. To fix this, we only retain
those entries in `cucumberFeaturesWithLineNumbers`, for which we can find a cor-
responding entry in the list of already filtered `_spec`s.