fix(wdio-cucumber-framework): filter cucumberFeaturesWithLineNumbers (v8) (#13563)
* test: correct cucumber feature file execution

* test: run cucumber feature with line number only once

* test: add dummy scenario to verify line number execution

If the dummy scenario were to ever execute, it would break the
previously introduced test logic.

* test: move "once"-check into assertExecutingFeatureFile

This simplifies the smoke test a lot.

* fix(wdio-cucumber-framework): filter cucumberFeaturesWithLineNumbers (webdriverio#13560)

While `_specs` has already been filtered here, `cucumberFeaturesWithLineNumbers`,
which just exists to retain the line numbers, still contains all specs as passed
on the cmdline or in the config. Therefore, just passing these to cucumber will
inevitably execute all feature lines in every worker. To fix this, we only retain
those entries in `cucumberFeaturesWithLineNumbers`, for which we can find a cor-
responding entry in the list of already filtered `_spec`s.