🐛⏪️ revert: "tsConfigPath in wdio.conf (#14664)" (#14802)
* 🐛⏪️ revert: "tsConfigPath in wdio.conf (#14664)"

This reverts most of commit 6173b79c65637c2300930533778f245521a14883.

The introduction of jiti caused significant regressions for users with complex `wdio.conf.ts` files, as reported in #14801. This change restores the previous import mechanism to resolve these issues.

It retains the attempt to load the config file and parse the tsConfigPath setting from it, with a cache-busting mechanism to prevent Node from caching failed imports. Documentation has been updated to explain the limitations of this configuration method.

Fixes #14801

* ✅ Resolve vitest "Please remember to await the assertion" warnings

An example of the warning in the test logs:
```
Promise returned by `expect(actual).rejects.toThrowError(expected)` was not awaited. Vitest currently auto-awaits hanging assertions at the end of the test, but this will cause the test to fail in Vitest 3. Please remember to await the assertion.
    at /home/steve/dev/webdriverio/packages/wdio-shared-store-service/tests/client.test.ts:38:43
```

* ✅ Update mock path to match changes in https://github.com/webdriverio/guinea-pig/pull/9

* ✅ Attempt to avert flaky tests by increasing the timeout.

* ✅ limit the concurrent instances to one less than the number of CPUs

* ✅ use `beforeEach` rather than `before` to give the tests a known starting state

* ✅ disable infobars in the browser so that the viewport size is more predictable.

* ✅ await async onComplete in isRunning test